Fixed some size_t/int warnings. Fixed compilation error on Windows for ExtractObject tool (include pcl_ros/transform.h missing).
git-svn-id: http://rtabmap.googlecode.com/svn/trunk/rtabmap@1295 f169173b-cf89-36c8-b27e-44dbe73f0c83

+++ corelib/include/rtabmap/core/Transform.h
@@ -38,7 +38,7 @@ public:
 
 	const float * data() const {return data_.data();}
 	float * data() {return data_.data();}
-	int size() const {return data_.size();}
+	int size() const {return (int)data_.size();}
 
 	float & x() {return data_[3];}
 	float & y() {return data_[7];}
